Added Bonus Apple
I have added code that spawns a bonus blue apple every 5 normal apples.
If eaten it adds 5 to the score, if you eat the red apple first,
however, the bonus apple disappears.

I've implemented this by declaring a new set of coords - gBonusApple,
and setting it up in the exact same way as gApple. I've also declared a
global int gBonusCounter that is ++'d after every eaten bonus, this is
reset to 0 when the player loses the game.

The bonus apple is only spawned when a new function, bonusIsSpawned()
returns true. It does so when the score % 5 == 0.

Any other questions, or something I've missed in this explanation,
message on the chat muthafuckas.

Peace